+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - PR #1124 Sub-communicator initialization for 2D partitioning support
 - PR #838 Add pattern accelerator API functions and pattern accelerator API based implementations of PageRank, Katz Centrality, BFS, and SSSP
 - PR #1147 Added support for NetworkX graphs as input type
+- PR #1157 Louvain API update to use graph_container_t
 
 ## Improvements
 - PR 1081 MNMG Renumbering - sort partitions by degree
@@ -18,6 +19,7 @@
 - PR #1139 MNMG Louvain Python updates, Cython cleanup
 - PR #1149 Parquet read and concat within workers
 - PR #1152 graph container cleanup, added arg for instantiating legacy types and switch statements to factory function
+- PR #1162 enhanced networkx testing
 
 ## Bug Fixes
 - PR #1131 Show style checker errors with set +e